+++ src/main/kotlin/com/coder/toolbox/CoderToolboxContext.kt
@@ -1,7 +1,9 @@
 package com.coder.toolbox
 
+import com.coder.toolbox.settings.SettingSource
 import com.coder.toolbox.store.CoderSecretsStore
 import com.coder.toolbox.store.CoderSettingsStore
+import com.coder.toolbox.util.toURL
 import com.jetbrains.toolbox.api.core.diagnostics.Logger
 import com.jetbrains.toolbox.api.localization.LocalizableStringFactory
 import com.jetbrains.toolbox.api.remoteDev.connection.ClientHelper
@@ -20,4 +22,42 @@ data class CoderToolboxContext(
     val i18n: LocalizableStringFactory,
     val settingsStore: CoderSettingsStore,
     val secrets: CoderSecretsStore
-)
+) {
+    /**
+     * Try to find a URL.
+     *
+     * In order of preference:
+     *
+     * 1. Last used URL.
+     * 2. URL in settings.
+     * 3. CODER_URL.
+     * 4. URL in global cli config.
+     */
+    val deploymentUrl: Pair<String, SettingSource>? = this.secrets.lastDeploymentURL.let {
+        if (it.isNotBlank()) {
+            it to SettingSource.LAST_USED
+        } else {
+            this.settingsStore.defaultURL()
+        }
+    }
+
+    /**
+     * Try to find a token.
+     *
+     * Order of preference:
+     *
+     * 1. Last used token, if it was for this deployment.
+     * 2. Token on disk for this deployment.
+     * 3. Global token for Coder, if it matches the deployment.
+     */
+    fun getToken(deploymentURL: String?): Pair<String, SettingSource>? = this.secrets.lastToken.let {
+        if (it.isNotBlank() && this.secrets.lastDeploymentURL == deploymentURL) {
+            it to SettingSource.LAST_USED
+        } else {
+            if (deploymentURL != null) {
+                this.settingsStore.token(deploymentURL.toURL())
+            } else null
+        }
+    }
+
+}